Fix selection bug in firefox when double-click an element. Adding new 'modify' tab to the toolbar, which opens automatically, when a modify action for the select tag exists.
Take a look in the MooEditable.Table class to see, how the modify
actions works. Fix small webkit bug when selection a image
(selectNode()). New option in a action 'modify', to handle actions for
the 'modify' tab. Calculate very poor the new position of the flying
toolbar (Need work here)